MOD: move files into src folder
