+++ src/lib/policy/sign_up.js
@@ -1,16 +1,14 @@
-const {
-	interactionPolicy: { Prompt, Check }
-} = require('oidc-provider');
+import { interactionPolicy } from 'oidc-provider';
 
-module.exports = interactions => {
-	const signUp = new Prompt({
+export default (interactions) => {
+	const signUp = new interactionPolicy.Prompt({
 		name: 'sign_up',
-		requestable: true
+		requestable: true,
 	});
 
 	signUp.checks.remove('sign_up_prompt');
 	signUp.checks.add(
-		new Check('sign_up_prompt', 'Sign up prompt was not resolved', '', ctx => {
+		new interactionPolicy.Check('sign_up_prompt', 'Sign up prompt was not resolved', '', (ctx) => {
 			const { oidc } = ctx;
 			// If Sign Up is already in-progress or was requested for by a client
 			if (oidc.prompts.has('sign_up') && oidc.promptPending('sign_up')) {
